// PolicyManagementAngular/src/app/app-entites/models/customer/add-update-customer-model.ts
export interface IAddUpdateCustomerModel {
    Id: number;
    BranchId: number;
    CustomerNameSalutation: number;
    CustomerName: string;
    CustomerType: number;
    CustomerContactSalutation: number;
    CustomerContact: string;
    DecisionMaker: boolean;
    CustomerCluster: number;
    CompanyTerritory: number;
    Address1: string;
    City1: number;
    Pincode1: string;
    Gstin1: string;
    Address2: string;
    City2: number;
    Pincode2: string;
    Gstin2: string;
    Address3: string;
    City3: number;
    Pincode3: string;
    Gstin3: string;
    Mobile1: string;
    Mobile2: string;
    Phone1: string;
    Phone2: string;
    Email1: string;
    Email2: string;
    CommunicationOptOut1: boolean;
    CommunicationOptOut2: boolean;
    CommunicationOptOut3: boolean;
    CommunicationOptOut4: boolean;
    SelectedPolicyAddress: number;
    SelectedWhatsAppCommunication: number;
    SelectedMobileCommunication: number;
    ReferBy: number;
    Pos: number;
    TeamMember: number;
    Reference: number;
    Pan: string;
    DateOfBirth: string;
    DateOfAnniversary: string;
    Aadhaar: string;
    MaritalStatus: number;
    NumberOfDependent: number;
    Profession: number;
    LineOfBusiness: number;
    Industry: number;
    Designation: number;
    IsPos: boolean;
    IsTeamMember: boolean;
    PassportNumber: string;
    Gender: number;
}